Database refactor for separating the candidate and cv parts is complete.

+++ src/main/java/org/smartjobs/core/service/CreditService.java
@@ -3,7 +3,7 @@ package org.smartjobs.core.service;
 public interface CreditService {
     boolean userHasEnoughCredits(long userId);
 
-    int userCredit(long userId);
+    long userCredit(long userId);
 
     boolean debitAndVerify(long userId, int size);
 }
